requirements.txt
a2wsgi==1.10.10
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.1.0
apache-airflow-core==3.1.0
apache-airflow-providers-common-compat==1.7.4
apache-airflow-providers-common-io==1.6.3
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-google==18.0.0
apache-airflow-providers-smtp==2.3.1
apache-airflow-providers-standard==1.9.0
apache-airflow-task-sdk==1.1.0
appnope==0.1.4
argcomplete==3.6.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.10.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.14.2
bleach==6.2.0
cachetools==6.2.1
cadwyn==5.6.0
cattrs==25.3.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
cycler==0.12.1
db-dtypes==1.4.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
dotenv==0.9.9
email-validator==2.3.0
executing==2.2.1
fastapi==0.119.1
fastapi-cli==0.0.14
fastjsonschema==2.21.2
fastuuid==0.14.0
filelock==3.20.0
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
gcloud==0.18.3
gcloud-aio-auth==5.4.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.6.0
gcsfs==2025.9.0
geographiclib==2.1
geopy==2.4.1
google-ads==28.3.0
google-analytics-admin==0.26.0
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.122.0
google-cloud-alloydb==0.6.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-automl==2.17.0
google-cloud-batch==0.18.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-datatransfer==3.20.0
google-cloud-bigtable==2.34.0
google-cloud-build==3.33.0
google-cloud-compute==1.40.0
google-cloud-container==2.61.0
google-cloud-core==2.4.3
google-cloud-datacatalog==3.28.0
google-cloud-dataflow-client==0.10.0
google-cloud-dataform==0.7.0
google-cloud-dataplex==2.13.0
google-cloud-dataproc==5.23.0
google-cloud-dataproc-metastore==1.20.0
google-cloud-dlp==3.33.0
google-cloud-kms==3.7.0
google-cloud-language==2.18.0
google-cloud-logging==3.12.1
google-cloud-managedkafka==0.2.0
google-cloud-memcache==1.13.0
google-cloud-monitoring==2.28.0
google-cloud-orchestration-airflow==1.18.0
google-cloud-os-login==2.18.0
google-cloud-pubsub==2.31.1
google-cloud-redis==2.19.0
google-cloud-resource-manager==1.15.0
google-cloud-run==0.12.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==2.19.0
google-cloud-storage-transfer==1.18.0
google-cloud-tasks==2.20.0
google-cloud-texttospeech==2.33.0
google-cloud-translate==3.22.0
google-cloud-videointelligence==2.17.0
google-cloud-vision==3.11.0
google-cloud-workflows==1.19.0
google-crc32c==1.7.1
google-genai==1.46.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
greenback==1.2.1
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-gcp==0.2.2
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
immutabledict==4.2.2
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.3.0
lazy-object-proxy==1.12.0
libcst==1.8.5
linkify-it-py==2.0.3
litellm==1.76.3
lockfile==0.12.2
looker-sdk==25.16.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
methodtools==0.4.7
mistune==3.1.4
more-itertools==10.8.0
msgspec==0.19.0
multidict==6.7.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.3
oauth2client==4.1.3
oauthlib==3.3.1
openai==2.6.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandas-gbq==0.29.2
pandocfilters==1.5.1
parso==0.8.5
pathlib_abc==0.5.2
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydata-google-auth==1.9.1
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
PyYAML-ft==8.0.0
pyzmq==27.1.0
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-argparse==1.7.1
rich-toolkit==0.15.1
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
Send2Trash==1.8.3
setproctitle==1.3.7
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlalchemy-bigquery==1.15.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.17.1
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.48.0
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
types-protobuf==6.32.1.20250918
types-python-dateutil==2.9.0.20251008
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.3.4
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.14
wirerope==1.0.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0